refactor(agent): split manager runtime modules
