+++ src/agent/manager/runtime.rs
@@ -1,469 +1,17 @@
-use std::collections::HashMap;
 use std::path::Path;
-use std::sync::Arc;
-use std::time::Duration;
 
 use chrono::Utc;
-use sqlx::{Row, SqlitePool};
-use tokio::io::{AsyncBufReadExt, BufReader};
 use tokio::process::Command;
-use tokio::sync::{RwLock, broadcast};
-use uuid::Uuid;
 
-use super::codec::{is_acp_message, is_dir_empty};
+use super::codec::is_dir_empty;
+use super::worktree::{repo_find_worktree_entry, worktree_ref_matches};
 use super::{
-    AgentHandle, AgentInput, AgentManager, expand_tilde, normalize_agent_loop_config,
-    normalize_path, spawn_agent_loop_controller, worktree_mode_to_str,
+    AgentInput, AgentManager, expand_tilde, normalize_agent_loop_config,
+    spawn_agent_loop_controller, worktree_mode_to_str,
 };
-use crate::agent::event_message_codec::persist_agent_event;
-use crate::agent::{AgentOutput, AgentRecord, OutputStream, WorktreeMode};
-use crate::push::PushService;
-use agenthub_db::{AgentEventDbRouter, AgentEventIdleGc};
-
-#[derive(Debug, Clone, PartialEq, Eq)]
-struct GitWorktreeEntry {
-    path: String,
-    head: Option<String>,
-    branch: Option<String>,
-}
-
-async fn repo_find_worktree_entry(
-    repo: &str,
-    workdir: &str,
-) -> anyhow::Result<Option<GitWorktreeEntry>> {
-    let output = Command::new("git")
-        .arg("-C")
-        .arg(repo)
-        .arg("worktree")
-        .arg("list")
-        .arg("--porcelain")
-        .output()
-        .await?;
-    if !output.status.success() {
-        let stderr = String::from_utf8_lossy(&output.stderr);
-        let reason = stderr.trim();
-        if reason.is_empty() {
-            anyhow::bail!("git worktree list failed");
-        }
-        anyhow::bail!("git worktree list failed: {reason}");
-    }
-    let stdout = String::from_utf8_lossy(&output.stdout);
-    let target = normalize_worktree_path(workdir);
-    Ok(parse_worktree_list(&stdout)
-        .into_iter()
-        .find(|entry| normalize_worktree_path(&entry.path) == target))
-}
-
-fn parse_worktree_list(stdout: &str) -> Vec<GitWorktreeEntry> {
-    let mut entries = Vec::new();
-    let mut current: Option<GitWorktreeEntry> = None;
-    for line in stdout.lines() {
-        if line.is_empty() {
-            if let Some(entry) = current.take() {
-                entries.push(entry);
-            }
-            continue;
-        }
-        if let Some(path) = line.strip_prefix("worktree ") {
-            if let Some(entry) = current.take() {
-                entries.push(entry);
-            }
-            current = Some(GitWorktreeEntry {
-                path: path.to_string(),
-                head: None,
-                branch: None,
-            });
-            continue;
-        }
-        if let Some(head) = line.strip_prefix("HEAD ") {
-            if let Some(entry) = current.as_mut() {
-                entry.head = Some(head.to_string());
-            }
-            continue;
-        }
-        if let Some(branch) = line.strip_prefix("branch ")
-            && let Some(entry) = current.as_mut()
-        {
-            entry.branch = Some(branch.to_string());
-        }
-    }
-    if let Some(entry) = current.take() {
-        entries.push(entry);
-    }
-    entries
-}
-
-fn trim_ref_prefix(value: &str) -> &str {
-    value.strip_prefix("refs/heads/").unwrap_or(value)
-}
-
-fn is_hex_sha(value: &str) -> bool {
-    let len = value.len();
-    (7..=64).contains(&len) && value.chars().all(|ch| ch.is_ascii_hexdigit())
-}
-
-fn worktree_ref_matches(entry: &GitWorktreeEntry, expected_ref: &str) -> bool {
-    let expected = expected_ref.trim();
-    if expected.eq_ignore_ascii_case("HEAD") {
-        return true;
-    }
-
-    let expected_branch = trim_ref_prefix(expected);
-    if !expected_branch.is_empty()
-        && let Some(branch) = entry.branch.as_deref()
-        && trim_ref_prefix(branch) == expected_branch
-    {
-        return true;
-    }
-
-    if is_hex_sha(expected)
-        && let Some(head) = entry.head.as_deref()
-    {
-        return head.starts_with(expected);
-    }
-
-    false
-}
-
-fn normalize_worktree_path(path: &str) -> String {
-    let canonical = std::fs::canonicalize(path).or_else(|err| {
-        tracing::warn!(
-            path = %path,
-            error = %err,
-            "failed to canonicalize worktree path"
-        );
-        if Path::new(path).is_absolute() {
-            return Ok(Path::new(path).to_path_buf());
-        }
-        std::env::current_dir().map(|cwd| cwd.join(path))
-    });
-    let canonical = canonical
-        .map(|value| value.to_string_lossy().to_string())
-        .unwrap_or_else(|_| path.to_string());
-    normalize_path(&canonical)
-}
+use crate::agent::{AgentRecord, WorktreeMode};
 
 impl AgentManager {
-    pub(super) async fn emit_run_status(
-        &self,
-        output_tx: broadcast::Sender<AgentOutput>,
-        agent_id: String,
-        session_id: String,
-        status: &str,
-    ) {
-        let message = serde_json::json!({
-            "type": "run_status",
-            "status": status,
-            "session_id": session_id,
-        })
-        .to_string();
-        let seq = Uuid::now_v7().to_string();
-        let ts = Utc::now().timestamp();
-        let event_id = match persist_agent_event(
-            &self.event_dbs,
-            self.idle_gc.as_ref(),
-            &agent_id,
-            &session_id,
-            &seq,
-            ts,
-            &OutputStream::Acp,
-            &message,
-        )
-        .await
-        {
-            Ok(event_id) => event_id,
-            Err(err) => {
-                tracing::error!(
-                    agent_id = %agent_id,
-                    session_id = %session_id,
-                    status = %status,
-                    error = %err,
-                    "emit_run_status failed to persist event"
-                );
-                return;
-            }
-        };
-        let output = AgentOutput {
-            event_id,
-            agent_id: agent_id.clone(),
-            session_id: session_id.clone(),
-            seq: seq.clone(),
-            ts,
-            stream: OutputStream::Acp,
-            message: message.clone(),
-        };
-        let _ = output_tx.send(output);
-    }
-
-    pub(super) async fn spawn_output_reader<R>(
-        &self,
-        agent_id: String,
-        session_id: String,
-        stream: OutputStream,
-        reader: R,
-        output_tx: broadcast::Sender<AgentOutput>,
-        detect_acp_messages: bool,
-    ) where
-        R: tokio::io::AsyncRead + Unpin + Send + 'static,
-    {
-        let event_dbs = self.event_dbs.clone();
-        let idle_gc = self.idle_gc.clone();
-        tokio::spawn(async move {
-            let mut lines = BufReader::new(reader).lines();
-            while let Ok(Some(line)) = lines.next_line().await {
-                let is_acp = detect_acp_messages && is_acp_message(&line);
-                let stream = if is_acp {
-                    OutputStream::Acp
-                } else {
-                    stream.clone()
-                };
-                let seq = Uuid::now_v7().to_string();
-                let ts = Utc::now().timestamp();
-                let event_id = match persist_agent_event(
-                    &event_dbs,
-                    idle_gc.as_ref(),
-                    &agent_id,
-                    &session_id,
-                    &seq,
-                    ts,
-                    &stream,
-                    &line,
-                )
-                .await
-                {
-                    Ok(event_id) => event_id,
-                    Err(err) => {
-                        tracing::error!(
-                            agent_id = %agent_id,
-                            session_id = %session_id,
-                            stream = ?stream,
-                            error = %err,
-                            "spawn_output_reader failed to persist event"
-                        );
-                        continue;
-                    }
-                };
-                let output = AgentOutput {
-                    event_id,
-                    agent_id: agent_id.clone(),
-                    session_id: session_id.clone(),
-                    seq: seq.clone(),
-                    ts,
-                    stream: stream.clone(),
-                    message: line.clone(),
-                };
-                let _ = output_tx.send(output);
-            }
-        });
-    }
-
-    pub(super) async fn spawn_exit_watcher(&self, agent_id: String, session_id: String) {
-        let db = self.db.clone();
-        let event_dbs = self.event_dbs.clone();
-        let idle_gc = self.idle_gc.clone();
-        let inner = self.inner.clone();
-        let push = self.push.clone();
-        let agent_id_clone = agent_id.clone();
-        tokio::spawn(async move {
-            let child = {
-                let guard = inner.read().await;
-                guard.get(&agent_id).map(|h| h.child.clone())
-            };
-
-            if let Some(child_mutex) = child {
-                let success = loop {
-                    let poll_result = {
-                        let mut child_guard = child_mutex.lock().await;
-                        let child = match child_guard.as_mut() {
-                            Some(child) => child,
-                            None => return,
-                        };
-                        child.try_wait()
-                    };
-                    match poll_result {
-                        Ok(Some(status)) => break status.success(),
-                        Ok(None) => {
-                            tokio::time::sleep(Duration::from_millis(100)).await;
-                        }
-                        Err(err) => {
-                            tracing::warn!(
-                                agent_id = %agent_id_clone,
-                                session_id = %session_id,
-                                error = %err,
-                                "spawn_exit_watcher failed to poll child status"
-                            );
-                            break false;
-                        }
-                    }
-                };
-                Self::finalize_process_exit(
-                    &db,
-                    &event_dbs,
-                    idle_gc,
-                    &inner,
-                    &push,
-                    &agent_id_clone,
-                    &session_id,
-                    success,
-                )
-                .await;
-            }
-        });
-    }
-
-    #[allow(clippy::too_many_arguments)]
-    pub(super) async fn finalize_process_exit(
-        db: &SqlitePool,
-        event_dbs: &AgentEventDbRouter,
-        idle_gc: Option<AgentEventIdleGc>,
-        inner: &Arc<RwLock<HashMap<String, AgentHandle>>>,
-        push: &Arc<PushService>,
-        agent_id: &str,
-        session_id: &str,
-        success: bool,
-    ) {
-        let row = sqlx::query(
-            r#"
-            SELECT ended_at FROM agent_sessions WHERE id = ?1
-            "#,
-        )
-        .bind(session_id)
-        .fetch_optional(db)
-        .await;
-        let row = match row {
-            Ok(row) => row,
-            Err(err) => {
-                tracing::warn!(
-                    agent_id = %agent_id,
-                    session_id = %session_id,
-                    error = %err,
-                    "finalize_process_exit failed to read existing session state"
-                );
-                None
-            }
-        };
-        let ended_at: Option<i64> = row.map(|r| r.get("ended_at"));
-        if ended_at.is_some() {
-            let mut guard = inner.write().await;
-            guard.remove(agent_id);
-            if let Some(idle_gc) = &idle_gc {
-                idle_gc.remove_agent(agent_id).await;
-            }
-            return;
-        }
-
-        let now = Utc::now().timestamp();
-        let state = if success { "completed" } else { "failed" };
-        let status = if success { "stopped" } else { "failed" };
-        if let Err(err) = sqlx::query(
-            r#"
-            UPDATE agent_sessions
-            SET status = ?1, ended_at = ?2
-            WHERE id = ?3
-            "#,
-        )
-        .bind(state)
-        .bind(now)
-        .bind(session_id)
-        .execute(db)
-        .await
-        {
-            tracing::warn!(
-                agent_id = %agent_id,
-                session_id = %session_id,
-                status = %state,
-                error = %err,
-                "finalize_process_exit failed to update agent_sessions state"
-            );
-        }
-
-        if let Err(err) = sqlx::query(
-            r#"
-            UPDATE agents
-            SET status = ?1, updated_at = ?2
-            WHERE id = ?3
-            "#,
-        )
-        .bind(status)
-        .bind(now)
-        .bind(agent_id)
-        .execute(db)
-        .await
-        {
-            tracing::warn!(
-                agent_id = %agent_id,
-                session_id = %session_id,
-                status = %status,
-                error = %err,
-                "finalize_process_exit failed to update agents state"
-            );
-        }
-
-        let seq = Uuid::now_v7().to_string();
-        let ts = Utc::now().timestamp();
-        let message = serde_json::json!({
-            "type": "run_status",
-            "status": state,
-            "session_id": session_id,
-        })
-        .to_string();
-        let event_id = match persist_agent_event(
-            event_dbs,
-            None,
-            agent_id,
-            session_id,
-            &seq,
-            ts,
-            &OutputStream::Acp,
-            &message,
-        )
-        .await
-        {
-            Ok(event_id) => event_id,
-            Err(err) => {
-                tracing::error!(
-                    agent_id = %agent_id,
-                    session_id = %session_id,
-                    status = %state,
-                    error = %err,
-                    "finalize_process_exit failed to persist run-status event"
-                );
-                let mut guard = inner.write().await;
-                guard.remove(agent_id);
-                if let Some(idle_gc) = &idle_gc {
-                    idle_gc.remove_agent(agent_id).await;
-                }
-                return;
-            }
-        };
-        let output = AgentOutput {
-            event_id,
-            agent_id: agent_id.to_string(),
-            session_id: session_id.to_string(),
-            seq,
-            ts,
-            stream: OutputStream::Acp,
-            message,
-        };
-        if let Some(handle) = inner.read().await.get(agent_id) {
-            let _ = handle.output_tx.send(output);
-        }
-
-        if let Err(err) = push.notify_agent_completed(agent_id, session_id).await {
-            tracing::warn!(
-                agent_id = %agent_id,
-                session_id = %session_id,
-                error = %err,
-                "finalize_process_exit failed to emit push completion notification"
-            );
-        }
-        let mut guard = inner.write().await;
-        guard.remove(agent_id);
-        if let Some(idle_gc) = &idle_gc {
-            idle_gc.remove_agent(agent_id).await;
-        }
-    }
-
     #[tracing::instrument(
         skip(self, agent),
         fields(
@@ -996,12 +544,9 @@ impl AgentManager {
 
 #[cfg(test)]
 mod tests {
-    use super::{GitWorktreeEntry, is_hex_sha, parse_worktree_list, worktree_ref_matches};
-    use crate::agent::OutputStream;
     use chrono::Utc;
     use sqlx::{Row, SqlitePool};
-    use tokio::io::AsyncWriteExt;
-    use tokio::time::{Duration, timeout};
+    use tokio::time::Duration;
     use uuid::Uuid;
 
     async fn build_test_state_with_idle_gc() -> crate::state::AppState {
@@ -1096,146 +641,6 @@ mod tests {
             .await
             .insert(agent_id.clone(), handle);
         (agent_id, session_id)
-    }
-
-    async fn collect_streams_for_lines(detect_acp_messages: bool, lines: &[&str]) -> Vec<String> {
-        let state = crate::api::team_tests::build_test_state().await;
-        let (agent_id, session_id) = insert_agent_and_session(&state.db, "stream-classify").await;
-        let (mut writer, reader) = tokio::io::duplex(4096);
-        let (output_tx, mut output_rx) = tokio::sync::broadcast::channel((lines.len() + 1).max(1));
-
-        state
-            .agents
-            .spawn_output_reader(
-                agent_id.clone(),
-                session_id.clone(),
-                OutputStream::Stderr,
-                reader,
-                output_tx,
-                detect_acp_messages,
-            )
-            .await;
-
-        for line in lines {
-            writer.write_all(line.as_bytes()).await.expect("write line");
-            writer.write_all(b"\n").await.expect("write newline");
-        }
-        writer.shutdown().await.expect("shutdown writer");
-        drop(writer);
-
-        for _ in 0..lines.len() {
-            timeout(Duration::from_secs(2), output_rx.recv())
-                .await
-                .expect("timed out waiting for output reader to broadcast event")
-                .expect("broadcast channel closed unexpectedly");
-        }
-        let event_db = state
-            .agents
-            .event_dbs
-            .pool_for_agent(&agent_id)
-            .await
-            .expect("open per-agent event db");
-        sqlx::query(
-            r#"
-            SELECT stream
-            FROM agent_events
-            WHERE session_id = ?1
-            ORDER BY id ASC
-            "#,
-        )
-        .bind(session_id)
-        .fetch_all(&event_db)
-        .await
-        .expect("list persisted streams")
-        .into_iter()
-        .map(|row| row.get::<String, _>("stream"))
-        .collect()
-    }
-
-    #[test]
-    fn parse_worktree_list_extracts_entries() {
-        let stdout = r#"
-worktree /tmp/repo
-HEAD 0000000000000000000000000000000000000000
-branch refs/heads/main
-
-worktree /tmp/repo/worktrees/agent-a
-HEAD 1111111111111111111111111111111111111111
-branch refs/heads/agent-a
-"#;
-        let entries = parse_worktree_list(stdout);
-        assert_eq!(entries.len(), 2);
-        assert_eq!(entries[1].path, "/tmp/repo/worktrees/agent-a");
-        assert_eq!(entries[1].branch.as_deref(), Some("refs/heads/agent-a"));
-    }
-
-    #[test]
-    fn worktree_ref_matches_accepts_head() {
-        let entry = GitWorktreeEntry {
-            path: "/tmp/repo/worktrees/agent-a".to_string(),
-            head: Some("1111111111111111111111111111111111111111".to_string()),
-            branch: Some("refs/heads/agent-a".to_string()),
-        };
-        assert!(worktree_ref_matches(&entry, "HEAD"));
-    }
-
-    #[test]
-    fn worktree_ref_matches_accepts_matching_branch() {
-        let entry = GitWorktreeEntry {
-            path: "/tmp/repo/worktrees/agent-a".to_string(),
-            head: Some("1111111111111111111111111111111111111111".to_string()),
-            branch: Some("refs/heads/agent-a".to_string()),
-        };
-        assert!(worktree_ref_matches(&entry, "refs/heads/agent-a"));
-        assert!(worktree_ref_matches(&entry, "agent-a"));
-        assert!(!worktree_ref_matches(&entry, "agent-b"));
-    }
-
-    #[test]
-    fn worktree_ref_matches_accepts_matching_commit_prefix() {
-        let entry = GitWorktreeEntry {
-            path: "/tmp/repo/worktrees/agent-a".to_string(),
-            head: Some("1111111111111111111111111111111111111111".to_string()),
-            branch: None,
-        };
-        assert!(worktree_ref_matches(&entry, "1111111"));
-        assert!(!worktree_ref_matches(&entry, "2222222"));
-    }
-
-    #[test]
-    fn is_hex_sha_accepts_sha256_length() {
-        assert!(is_hex_sha(
-            "0123456789abcdef0123456789abcdef0123456789abcdef0123456789abcdef"
-        ));
-    }
-
-    #[tokio::test]
-    async fn spawn_output_reader_promotes_latest_codex_event_types_for_acp_agents() {
-        let streams = collect_streams_for_lines(
-            true,
-            &[
-                r#"{"type":"plan","steps":[{"title":"Investigate"}]}"#,
-                r#"{"type":"available_commands","commands":["/compact","/undo"]}"#,
-                r#"{"type":"current_mode","current_mode_id":"code"}"#,
-                r#"{"type":"run_status","status":"completed","session_id":"s-1"}"#,
-                "plain stderr line",
-            ],
-        )
-        .await;
-        assert_eq!(streams, vec!["acp", "acp", "acp", "acp", "stderr"]);
-    }
-
-    #[tokio::test]
-    async fn spawn_output_reader_keeps_stderr_for_non_acp_agents() {
-        let streams = collect_streams_for_lines(
-            false,
-            &[
-                r#"{"type":"plan","steps":[{"title":"Investigate"}]}"#,
-                r#"{"type":"run_status","status":"completed","session_id":"s-1"}"#,
-            ],
-        )
-        .await;
-        assert_eq!(streams, vec!["stderr", "stderr"]);
     }
 
     #[tokio::test]
@@ -1630,25 +1035,6 @@ branch refs/heads/agent-a
             .expect("fetch session row");
         assert_eq!(session_row.get::<String, _>("status"), "exited");
         assert!(session_row.get::<Option<i64>, _>("ended_at").is_some());
-    }
-
-    #[tokio::test]
-    async fn finalize_process_exit_tolerates_closed_pool() {
-        let state = crate::api::team_tests::build_test_state().await;
-        let (agent_id, session_id) = insert_agent_and_session(&state.db, "finalize-closed").await;
-        state.db.close().await;
-
-        super::AgentManager::finalize_process_exit(
-            &state.db,
-            &state.agents.event_dbs,
-            state.agents.idle_gc.clone(),
-            &state.agents.inner,
-            &state.push,
-            &agent_id,
-            &session_id,
-            false,
-        )
-        .await;
     }
 
     #[tokio::test]
